lambda_function/requirements.txt
requests==2.32.3
aws-lambda==2.1.3
line-bot-sdk==3.11.1